Squashed 'phydmslib/Bpp/bpp-seq/' changes from 6978099..b1c4497
b1c4497 generalize Pasta to any alphabet, and some tests to partially back-up this claim
6136d72 Merge branch 'master' of biopp.univ-montp2.fr:git/bpp-seq
575282b make it so Pasta does not inherit from Fasta
409e850 Lexicon alphabet
51ea722 Merge branch 'master' of biopp.univ-montp2.fr:/git/bpp-seq
b108466 Lexicon alphabet
8078ae5 Lexicon alphabet
6cfa079 Added missing include files.
bfe75f2 Revert to c++11 for "older" compilers.
897e4bb Removed debian packaging.
6050887 Removed last NO_VIRTUAL_COV conditions.
14fc880 Compiled with -std=c++14
1c62930 Merge branch 'c++0x'
c4e1d55 Faster getCompleteSites.
dd917e3 Updated code for c++0x.

git-subtree-dir: phydmslib/Bpp/bpp-seq
git-subtree-split: b1c4497c031da9e3c3a1931c5e07546bf9bd428c

+++ src/Bpp/Seq/Alphabet/WordAlphabet.h
@@ -88,7 +88,7 @@ public:
    * @param pAlpha The Pointer to the Alphabet to be used.
    * @param num the length of the words.
    */
-  WordAlphabet(const Alphabet* pAlpha, unsigned int num);
+  WordAlphabet(const Alphabet* pAlpha, size_t num);
 
   WordAlphabet(const WordAlphabet& bia) : AbstractAlphabet(bia), vAbsAlph_(bia.vAbsAlph_) {}
